ClientPoolTimer: Workaround for ClientT oddity
With commit 1dbe653d21a5aeee6f9cad348d6f5b3fa3a41d49 logging got introduced to
the client pool timer. The log messages are logging the ClientT object and
thereby implicitly requiring that `operator<<` is implemented for that type.
This change notes the oddity in the code and changes the type in the newly
added test so that it compiles.