TRS: Wait for last_block_id + 1 live update in SubscribeToUpdates
In the read-path of legacy events, if the requested block ID is
`last_block_id + 1`, where `last_block_id` is the last known block
ID in storage, TRS must wait on the live update queue until
`last_block_id + 1` gets added to the queue.
Before this change, TRS would wait on the live update queue for any
live update and move on to SyncAndSend. If TRS moves on to syncAndSend
before `last_block_id + 1` gets added to storage, it may try to
read a non-existant block ID. This commit fixes this issue.