Change log level in getDirectorySize api
Disk space monitoring thread tracks size of the database,
we use boost::filesystem::recursive_directory_iterator for this.
It enumerates all the files in folder and then calculate size.
If rocksdb delete a sst files in between for compaction then we may see an error
Changing log level of this event to WARN