BCStateTran: memset BlockIOContext::blockData to claim ownership
When allocating only, memory is not marked as used, and other processed
might capture it later on. Some time can pass until replica may enter
into St as a source or as a destination. We would like to make sure the
memory is owned on startup, since the amount of memory acquired here
is very large. Failing in allocation on startup is better then failing
later on.
Solution: memset all memory in BlockIOContext::blockData immediately
after allocation in order marked it by OS as used and reserved for ST.