ConcordClient: Hide implementation details
ConcordClient as a library should not expose internal API. This includes
anything within the TRC and the client pool. The only exception is the bft
client interface because it is supposed to be public API and there is no point
in masking its types because ConcordClient is only a wrapper with additional
functionality.
This change uses the "pointer to implementation" approach so that ConcordClient
API users don't need to depend on the client pool or TRC at compilation time.