Introduce metrics to EventService
This commit introduces metrics for EventService. The following
metrics have been added -
1. num_writes_per_second
2. update_processing_dur
3. write_dur
While the first metrics describes the number of updates written on
to the gRPC stream by the event service, the latter two metrics
speify the time taken by the event service to process an update and
write an update to the stream. Note, write_dur is a subset of
update_processing_dur. We need both these metrics to identify if
the blocking write call is a bottleneck in update processing.