KvbAppFilter: Downgrade log level in getValueFromTagTable
When reading event groups TRS tries to calculate next global
and external event group IDs to read via call to `getValueFromTagTable`.
It is expected that the corresponding values have not yet been
populated in the tag table, wherein `getValueFromTagTable` would
throw and log an error. While the error is now re-thrown in the
caller `readEventGroups` method, we must also not log an ERROR
as the behavior is expected. This commit therefore downgrades the log to
WARN.