KvbAppFilter: findGlobalEventGroupId fix pruning
With pruned event groups there is a scenario in which either the oldest private
or the oldest public event group id is set to 0 although the newest is not 0.
That means, all event groups for this tag (public or private) have been pruned.
The algorithm in `findGlobaleventGroupId` will search within private event
groups first and looks at public event groups thereafter. Before this change,
the function would query private event groups that got pruned already and hence
fail. However, as long as the requested external event group id is in storage
the function needs to succeed.
In this change, we make sure that we skip private event groups if all have been
pruned and calculate the current_* pointers to point to last private event
group. Now, we can determine the index into public event groups as before.