CMF C++: Allow string as serialization buffer
This change adds the ability to serialize to and deserialize from `std::string`.
The user should always prefer `std::vector<uint8_t>` over a string because it
more accurately describes serialization and when would we ever need the
terminating NULL-byte?
However, `std::string` can be used to stay compatible with libraries that use
this type for byte buffers. Currently, C++ doesn't allow us to convert between
those types without copying. Therefore, with this change we can use CMF with
those libraries without performance overhead. However, it is considered a
*workaround* because byte buffers should not be strings in the first place.