Fix for correct timeout calculation in state snapshot service
Normally for GRPC client sets a deadline for the request. If the request
takes more time than the deadline, then client times out.
As State Snapshot service is a GRPC service, it also calculates a
timeout based on the deadline set. But the deadline was calculated in
reverse manner and this PR is fixing that issue. There are 2 changes
with this PR:
1) The deadline calculations are correctly done.
2) In case there is no deadline, the default timeout will become 30s
3) The maximum deadline that can be set from client side is 10 mins
So if the deadline is past (due to clock skew) or if the deadline is not
set, we will assume 30 sec as timeout. Any deadline between 1 millisec
to 600000ms is considered as valid. If the deadline is invalid then
30s will be choosen as default deadline.