Added test suite for restart recovery. (#2298)
* Added test suite for restart recovery.

In this test suite we will perform scenarios with replica restarts
during different states of the system to verify correct recovery of
replicas. This test suite is not meant to be part of the CI, because
it requires running over longer periods. It might be used for
daily/weekly runs.
Currently we have 2 scenarios:
1. Restart a replica while the system is processing client requests
   to verify the restarted replica recovers correctly and the system
   returns to FastCommit path.
2. Perform multiple View Changes and when the replicas start to initiate
   View change restart 1 random non-Primary replica. Then after the view
   change is complete verify the system restores Fast Commit path.

* Stabilize test to handle intermittent failures not representing problems.

Getting of the system's current view made more tolerant to delays, because
of the view change loop scenario.
Waiting for State Transfer takes in account the cases where replicas are half
Working Window apart.

* Added check for cases where we get View Change after State Transfer.

* Added description of the steps in the current restart/recovery scenarios.

* Added test case for View Change with the next primary down.

In the newly added test we crash both the current and the next
primaries to get the system in a view change with more than 1
View increments. In this state we restart one randomly selected
replica to verify correct restart recovering.

* Fix for test timeout failure.

We need to not restart the replica that will step in as primary,
because an additional View Change can be triggered, which we
don't handle in this test scenario.

* Simplify test_restarting_replica_with_client_load and fix quick checks.

Fix the quick checks for transition to slow/fast path and restructure
code for better control of the flow - now we have a check for transition
to slow path when we stop the selected replica and a check for transition
to fast path when we bring it back up again.

* Fix for performing complete cycles.

In order to have the test stable we need to perform full cicles of
stopping/bringing back up the selected replica. Using move_on_after
can interrupt in arbitrary stage.