Added restart recovery test for View Change with F crashed replicas. (#2611)
In this test we initiate View Change multiple times by restarting F replicas
including the current Primary, but excluding the next one. In this way
we perform a View Change with only 2F+1 replicas and in general we expect
only one View increment to happen per View Change, even though if a replica
initiates State Transfer this is not necessarily guaranteed and the test is
able to handle such scenarios correctly.