EventService: Handle async done notify
Only the active "event chain" which handles subscription and writing to the
stream should be able to unsubscribe and delete itself. The async notification
which comes from the gRPC server, not from the state machine, will set a flag
so that we stop the updates and finish (unsubscribe and delete) the stream.