Deactivate some thin_replica_server_test.cpp cases
This commit deactivates certain test cases in
thin-replica-server/test/thin_replica_server_test.cpp known to be
unstable. This is intended to be a temporary measure in order to prevent
these failures from blocking other merges, especially others intended to
improve stability. The intention is to reactivate these test cases in
the near future once a better fix for them can be provided.