Fix Preprepare sizing issues (#2649)
* **Problem Overview**  
MessageBase considers the span size outside of the regular "size" parameter. Hence, in some cases, we may end up letting a message be larger than the allowed size.  For example, if a prePrepare message span size is smaller than the max span size, we may end up with a message that exceeds a bit from the maximal size. This change is for fixing this issue
* **Testing Done**  
  CI tests