GrpcConnection: Re-use subscription thread
Continuously reading updates from subscription streams is the common case and
for some application performance critical. Therefore, instead of spawning a new
thread for every new Read() operation, this change introduces a thread pool
with a single thread to avoid the overhead.
This change has been tested as part of a wider effort to improve TRC.