Handoff: introduce an option to block and to push by priority, add tests
1) Introduce class PriorityQueue -
This template class implements a simple priority queue. It supports up
to <S> priorities.
Elements of type T are pushed into the queue with a given priority by
calling push(...).
To check the next element which is going to be popped, without popping
it, the user should call front().
Higher (larger) priority elements are processed before lower priority
elements. Elements with the same priority are
treated FIFO.
To get the total number of elements in the priority queue call size().
To check if the priority queue is empty call empty().

Comments:
a) std::priority_queue is not the same. Working with a priority_queue is
 similar to working with a heap.
It will not be efficient to implement a ?Compare function which makes
sure that the elements are popped in the exact
same order they were inserted, for any two elements with an equal
priority.
b) Calling pop() and front() on an empty queue is treated as an error
and throws a runtime_error exception.

2) Handoff: add an option for synchronous mode-  push and wait. This is
done by using class PriorityQueue. Blocked calls are priority 1,
non-blocked calls are priority 0.

3) Modify test multithreading.cpp to test new above features.