ReadOnlyReplica: new config param sendAskForCheckpointMsgPeriodSec
Via dynamic bft configuration: we set a new configuration parameter
sendAskForCheckpointMsgPeriodSec. This is the time in seconds to
periodically send sendAskForCheckpoint Message. This parameter
initializes askForCheckpointMsgTimer_. Till this point, we initialized
the timer with hard coded 5 seconds.

The 2nd change is the default: 5 seconds seems to be too little, and
overload the network with too many messages to start an ST cycle.
Therefore, we change this parameter to 30 by default, and probably
the same value will be set in the configuration file.