Fix for assertion on creating same checkpoint twice. (#2653)
Because we store the checkpoint in one transaction and update
the sequence ID in another we can potentially recover in a situation
where we have already created a checkpoint for the current sequence ID.
We need to resolve the situation in the recovery procedure, because
the storage mechanisms used in ReplicaImp and State Transfer are different
in order to improve decoupling and stacking the 2 transactions in 1 is not
feasible. This bug was found by the Apollo's restart recovery suite test
test_restarting_replica_with_client_load.