RVBManager: fix a bug in updateRvbDataDuringCheckpoint
1) Reverse order of RVT update: first prune the blocks, then add blocks.
both are optional of course. The reason is that the whole logic checks
the RVT min/max which should be most updated before adding blocks (
  in rare cases) for the most simplified code. Therefore, we would like
  to first trim the RVT.
2) Calculate add_range_min_block_id while taking into consideration
the pruned_blocks_digests_. We never want to define a range for blocks
that were already pruned.

Testing: test bkpCheckpointingWithPruning is added as a unit test to
reproduce the exact scenario for this fix. After the fix is applied,
the test won't fail.