Introduce Cppcheck in concord-bft dev image
Here we firstly introduce Cppcheck (2.8), a static analysis tool:
https://cppcheck.sourceforge.io/
https://github.com/danmar/cppcheck
https://en.wikipedia.org/wiki/Cppcheck

"Cppcheck is a static code analysis tool for the C and C++ programming
languages. It is a versatile tool that can check non-standard code.[2]
The creator and lead developer is Daniel Marjamäki. Cppcheck is free
software under the GNU General Public License."

Clang-tidy is a more powerful tool. On the other hand, there is quite
a wide consensus that using 2-3 tools for static analysis is almost
always better, as different issues may arise. Cppcheck is one of the
oldest and most widely used open-source tools for c++ static analysis
(I would say it is among the top 5).

At this stage, Cppcheck is not enforced in CI. It is part of the image,
with version 2.8 installed, and part of Makefile command 'make cppcheck'
which allows for manual check of the whole source code under
concord-bft, or part of it by defining a relative path or file(s).

There are 2 ways to run cppcheck:
1) Native or from inside a docker container. define CPPCHECK=ON and
generate. In that case, For every compiled file, cppcheck is triggered.
As of now, exit code is 0 and we do not fail the build.
2) Dockerized way - via concord-bft/Makefile: 'make cppcheck' to scan
the whole codebase. type 'make help' to see additional options. In this
case the script scripts/run-cppcheck.sh is triggered and handle the run.
One can also choose to run the script from inside the container. type
'run-cppcheck --help' for options.

For both ways, if there is output full_report.txt, the above script can
be used to analyze the results.

Configuration files, under .cppcheck
1) exitcode-suppressions.txt - currently empty, later can be used
when certain messages should be displayed but should not cause a
non-zero exit code.
2) includes.txt - used when only part of the codebase is analyzed.
3) suppressions.txt - Suppress warnings are listed in the file. Each
suppression is in the same format as <spec> above.