Added test with all non-primaries being isolated from each other. (#2720)
* Added test with all non-primaries being isolated from each other.

In this test we isolate the non-primary replicas form each other,
leaving only their connection to the primary. In this setup only
the primary is the only one who can bidirectionally communicate
with each replica. In this state the system can progress only up
to the working window size. We advance past the first stable checkpoint
which only the primary will collect and we initiate view change.
We perform this test in multiple cycles to verify correct recovery
and in the end we drop all adversaries and verify fast path is
recovered.

* Added step by step test description.