Added lastStableCheckpoint and proofForLastStable to VCMsg.
Now that we have modeled Checkpoint generation in our system
we are ready to model its implications on the View Change process
and only insert the relevant SequenceID-s to the View Change message
we generate. Also added clearing of the Working Window once we are
ready to activate the next View.