Fix in Replica.LiteInv regarding the stored newViewMsg-s
Since we store NewViewMsg-s in a set without discarding old ones
we need to check that they are valid for the View they have been
used for. We only use the one for the View we are in though, therefore
further in the model's development we might discard old ones.