Refactor recursive invocation of Message.valid predicate.
We had a recursive invocation of the valid member predicate
in Message when called on a New View msg, since this message
contains in itself a set of 2F+1 View Change msg-s, we need to
check their validity as well. This was causing problem with
Dafny's heuristics about recursion expansion. By refactoring
into separate predicates the check on New View msg validity
and the View Change msg validity we can resolve the problem
by explicitly calling validViewChangeMsg on the VC msgs contained
in the New View msg and avoiding the recursive call.