Reducing the severity of log message in CRE
A replica in state transfer continuously polls for reconfiguration
update from other non-ST replicas. If there is a loss of quorum due to
view change or any other reason, state transfer cre client won't be able
to get the updates from replica. We are reducing severity of this
message as this is a side effect of a more critical problem i.e. loss of
system's liveness.