Added state to Checkpoint messages.
Currently we use the map containing the CommittedClientOperations as
the state that we will transfer and verify. Later in the model development
we will have an additional set - ExecutedClientOperations that will be used
instead. We are keeping this change for the future to keep the model simple
for the first version of the proof and build up incrementally.
Also added a PerformStateTransfer step that slides the Working Window
further than the Working Window size.