Introducing an implementation for the UTT signature collector (#2804)
n this PR we introduce the basic implementation of the UTT signature processor component.
The signature processor is responsible for computing the partial signature, collecting others' partial signatures, assembling the full signature, and publishing it back through a consensus to the system.

In this implementation, we introduce a leader-based approach that relies on the bft consensus for correctness and safety.
Also, the signature processor is thread-safe such that the entry point method can be called by any thread.

When handling failures (i.e. the full valid signature was not received by consensus in time) we go over the jobs and retransmit every overdue job to all. However, if there are many jobs this may take a while. For that, we also save the partial signatures in an ordered map (ordered by the overdue date of the job) which allows us to go only over the jobs which exceeded their timeout.

Testing Done
We also added some simple unittests that demonstrate how to use the signature processor