Added State Transfer implementation.
Currently the state of each replica is the ordered client requests.
In the Checkpoint messages each replica puts its current state.
When a stable checkpoint is collected and it is for a state beyond
the state of the replica that collected it state transfer can be performed
by taking the agreed state.