Added predicate for IsIntraViewMsg.
In SBFT there are messages that are only relevant within a given active View.
those are currenly PrePrepare Prepare and Commit. We need to add predicates to
the Invariant that preserve the knowledge that replicas will not send those
messages beyond the boundaries of their Working Window and will not send
such message for a View higher than the last one they have activated.
More debug code added.