Formal modeling of view change + checkpointing (#2758)
* Introducing back the View Change steps.

We are bringing back the View Change steps now
that we have modeled the Checkpointing, which has
implication on the View Change process.

* Added lastStableCheckpoint and proofForLastStable to VCMsg.

Now that we have modeled Checkpoint generation in our system
we are ready to model its implications on the View Change process
and only insert the relevant SequenceID-s to the View Change message
we generate. Also added clearing of the Working Window once we are
ready to activate the next View.

* Valid NewViewMsg-s hold ViewChangeMsg-s in payload.vcMsgs.msgs

For some reason Dafny cannot see that the valid predicate of the
NewViewMsg does this check when using the predicate of the datatype.
When using a standalone predicate with the same check the verification
passes. Will look into the syntax of Dafny further.

* Fix in Replica.LiteInv regarding the stored newViewMsg-s

Since we store NewViewMsg-s in a set without discarding old ones
we need to check that they are valid for the View they have been
used for. We only use the one for the View we are in though, therefore
further in the model's development we might discard old ones.

* Clear the working window upon leaving the View.

We clear the working window on leaving the current View,
after we have transferred the relevant information in the
ViewChangeMsg we will send. We set the correct lastStableCheckpoint
Once we receive the NewViewMsg, which contains all necessary information.

* When we receive a View Change msg we also need to check the Checkpoint msgs signatures.

* When we leave a View we can only go to a higher View.

* Refactor recursive invocation of Message.valid predicate.

We had a recursive invocation of the valid member predicate
in Message when called on a New View msg, since this message
contains in itself a set of 2F+1 View Change msg-s, we need to
check their validity as well. This was causing problem with
Dafny's heuristics about recursion expansion. By refactoring
into separate predicates the check on New View msg validity
and the View Change msg validity we can resolve the problem
by explicitly calling validViewChangeMsg on the VC msgs contained
in the New View msg and avoiding the recursive call.

* Code refactoring regarding readability:

1. Created a predicate UniqueSenders that checks whether
   all messages in a set are from different senders. It is
   now used in all necessary places in messages.dfy
2. committedClientOperations is taken out of the Working Window
   and placed in the variables of the Replica. Also it no longer
   holds Option as the value associated with each key. Now that it
   is a finite map, the things that have not yet been committed
   simply are not present in the map.
3. Clearing the Working Window now happens in one step - EmptyWorkingWindow.
   This function takes care of clearing the votes we no longer need after
   View Change is initiated.

* Added state to Checkpoint messages.

Currently we use the map containing the CommittedClientOperations as
the state that we will transfer and verify. Later in the model development
we will have an additional set - ExecutedClientOperations that will be used
instead. We are keeping this change for the future to keep the model simple
for the first version of the proof and build up incrementally.
Also added a PerformStateTransfer step that slides the Working Window
further than the Working Window size.

* Readability changes and better structuring of comments.

* Added State Transfer implementation.

Currently the state of each replica is the ordered client requests.
In the Checkpoint messages each replica puts its current state.
When a stable checkpoint is collected and it is for a state beyond
the state of the replica that collected it state transfer can be performed
by taking the agreed state.

* Changed invariant predicate.

Now that replicas can perform garbage collection when moving
the Working Window, we need to base the invariant on the messages
in the network, since the network is our arbiter that keeps all
messages indefinitely.

* Added predicate EverySentCommitIsInWorkingWindowOrBefore to Inv.

Now that we have finite Working Window we need to preserve the knowledge
that honest replicas will not send a Commit beyond the highest Sequence ID
they have available in their Working Window.

* Added performance optimizations.

We add triggering hints to dafny's heuristics to avoid
unnecessary instantiations.

* Primary's PrePrepare needs to be in Active Window.

An honest primary will only send a PrePrepare that is
within its Working window and for the View it is in. The primary will
lock on it (will not send a different - conflicting PrePrepare).

* Added predicate for IsIntraViewMsg.

In SBFT there are messages that are only relevant within a given active View.
those are currenly PrePrepare Prepare and Commit. We need to add predicates to
the Invariant that preserve the knowledge that replicas will not send those
messages beyond the boundaries of their Working Window and will not send
such message for a View higher than the last one they have activated.
More debug code added.

* Added predicate EveryPrepareClientOpMatchesRecordedPrePrepare to Inv.

With this predicate added to the Invariant we preserve the knowledge
that replicas vote with Prepare only for the first PrePrepare they have
received and stored from the Primary.
Debugging code removed as the proof now passes correctly.