Check request service's health status when updating aggregate health
Previously, the health status reported for aggregate health queries,
i.e., queries which didn't specify any service name, was the last
known aggregate health status.
With this change, when queried for aggregate health, request service's
health will be checked in real time to retrieve the most up-to-date
health status of the service.