kVBC v4_blockchain: refactor multiple functions in KVBc, fix bugs 
KeyValueBlockchain::parentDigest and
KeyValueBlockchain::calculateBlockDigest return an zeroed digest if the
block is not found.
That doesn't make sense since it corrupts the outer state caller logic. 
The caller thinks it got the block and continues as usual. There is no
place stating that a zero digest is assigned in error. 
Using the API as a comment is still a not the best solution, because the a
zero digest, like all others, is a valid digest. 
Other functions (see ST blockchain) assert if block does not exist.They
should all return an empty optional like all others.

Fix the bug on error handling and return a null value as an option if
the block is not found. 
This is also done in multiple functions to prevent future bugs.
Unittests adjust accordingly.