Measure the number of storage reads made via TRS and KvbAppFilter (#2838)
* TRS: Add metric to count number of reads from storage

This commit adds a counter type metric that counts the number of
read requests made to the storage layer via TRS directly and
KvbAppFilter used by the TRS.
This is to track read activity triggered on the storage layer
via TRS and kvb_filter.

* Add logs to print storage reads via newest_public_event_group_record_time.h

The method `newestPublicEventGroupRecordTime` is used to get the time of
the last application-level transaction stored in the blockchain. It uses the
kvb_filter to get the newest public event group from storage.
This commit adds a log to print the number of reads made by this component.