Changes for building the config-gen tool in mac (#2779)
To support the internal tool in mac native environment, we need to include new directory wrt to mac and it should only hit only in case of mac compilation, so added the necessary changes under a flag(BUILD_CONFIG_GEN_TOOL_FOR_MAC). The changes in this PR will not have any effect when we compile the code in linux environment.