ReplicaImp, BCStateTran: fix race conditions on State Transfer state.
The way ST is reporting on completion is not thread-safe.
Here we introduce a new atomic flag, externalCollectingState_,
which is used to mark the ST state.
When ST finishes all its operations, it sets the flag to false.
notify the outside (which uses polling). The reporting is done when last
guard is destroyed and all callbacks are invoked, including Consensus
callback. So there is no change in that state during this time (which
caused a race condition).
ReplicaImp won't use the local flag anymore and won't call ST.
dataStore. It polls "isCollectingState" with the already existing API.
isCollectingState().

Here we introduce an "external state" which is polled by any external 
entity (e.g consensus) by calling API isCollectingState().
External state is changed only when ST is 100% sure about it's state,
All guards are destroyed, and dataStore reflects the new state.
In addition, the 2nd group of callbacks is  called only after the
external state is false.