Fix for halting CRE activity after ST. (#2882)
* Fix for halting CRE activity after ST.

We need to halt the CRE activity independently of the
presence of reconfiguration events once State Transfer
completes. Otherwise the polling loop never terminates
and leads to unnecessary read load on the system.

* Reduce the intensity of CRE polling.

Set the value of initial_retry_timeout, min_retry_timeout
and max_retry_timeout to 1s thus keeping the polling to a
steady non intrusive phase.