Introduce Multi size buffer pool (MultiSizeBufferPool)
This pool replaces RawMemoryPool. It was was mainly done due to the
very large internal fragmentation while allocating from single pool,
and to support a retry when allocated size is insufficient.
Definitions:
- Buffer: a continuous, fixed size array of bytes dedicated to a caller.
- Chunk: Wraps a buffer.Pool deals internally with chunks, which include
header and trailer metadata for internal
use.
- Used buffer/chunk: currently used by caller
- Unused buffer/chunk: currently not-used by caller, and stored in the
pool.
- Allocated buffer/chunk: currently allocated by pool. Can be used or
unused in that case.
- Non-allocated buffer/chunk: currently not allocated by pool - pool
reserves the right to allocate this chunk
dynamically.
- Subpool: holds a group of buffers of a fixed constant size. Each
subpool has its own configuration and contention.
lock-free domain: threads that try to get a buffer from subpool A, do
not compete with threads that try to get a
buffer from subpool B. This allow (with an additional very small
implementation changes) to create multiple
subpools with the same buffer size to reduce contention. Currently, all
subpools should be different sizes.
- Client - a user of the pool's API in order to use the memory pool
buffer.
- IPurger: a module that is responsible for purging the sub-queues,
based on recently collected data. Can be
disabled. Can be passed externally by pool creator. If not passed
externally, the pool creates a default
DefaultPurger. The main idea of the purger is to deallocate unused
resources (chunks) from the pool to keep it
minimal according to a given purging policy. Purger can be configured to
 treat different sub-pools (buffer sizes)
in different ways.
- ISubpoolSelector: a module that is responsible for selecting the most
suitable sub-pool (buffer size), based on
recently collected data. Can be disabled. It can be passed externally
by the pool creator. If not passed externally,
pool creates a default DefaultSubpoolSelector which chooses to smallest
buffer size sub-pool always.
- Pool - holds its configuration purger, subpool selector and an array
of subpools (at least 1).
Overview:
This memory pool has the following attributes:
- Thread safe for all of its operations (not thread safe, of course,
when accessing the currently used buffers).
- Supports advanced configuration, in particular caller can define any
amount of sub-pools of any size with initial
and maximal number of buffers for each subpool.
A caller may control the overall size of the allocated bytes in the
pool, this can give more 'freedom' per subpool,
and restrict the total pool size.
- Rich statistics:
1) Histograms (snapshots) reported to logs. Currently, histograms are
for this pool can't be views
from concord-ctl.
2) Metrics - reported to Wavefront.
3) Statistics - reported to log. It is very important to mention, that
 due to speed consideration, each of the
statistics are "stand-alone" and atomic. That means, for a statistics
report, you might find discrepancies between
statistics of the same type, due to the nature of this multi-threaded
pool (statistics are updated all the time by
multiple clients in a non-atomic way). All types of statistics can be
disabled/enabled.
- Blocking - calls to get a buffer should never fail, unless requested
buffer size is too big or there is no more
memory. - they always block. Currently, there is no implementation for
wait_for(..).
- Dynamic allocation - each pool allocates its initial chunks when
created. In case there is no available
chunks in a specific subpool, a new one gets dynamically allocated -
up to a maximum number of chunks defined for
this subpool.
- Throwing - asserts are done to detect internal coding errors. In
case of user input error or System error -
an exception is thrown.
- Borrowing - if all buffers of a certain sub-pool SP1 are allocated, a
calling thread checks for a free chunk in the
rest of the subpools (SP2, SP3 etc..) for speed, but won't try to
allocate a new one on these subpools. If a buffer
isn't found - it returns to wait for a chunk to be released by another
thread on the original subpool SP1.
- Unit-testing: done by GTest, see MultiSizeBufferPool_test.
- Safety checks:
1) Pool assign magic flags on chunk's header and trailer. If they are
 changed - exception is thrown.
2) Check for leaks - pool can be configured to check that all buffers
have been returned to it before destroying
itself. If not, it throws.
Locking is done only for conditional variables for this pool. This is
in order to achieve speed. This pool currently
faces thousands of calls per second for its main use-case. This leads
to sub-optimized cases in both statistics and
behavior. For example, the upper threshold for maximal pool size might
cross more than user defined, due to the
multi-threaded nature and since there is no locking . Another example
is for purge limit: we might purge a "little
more" than requested.