MultiSizeBufferPool: remove invalid thrown exception, avoid overflow (#2960)
Due to the thread model of the pool, which avoid as much as possible
from locking, I have to remove one of the exception thrown. The
exception logic is invalid in some rare cases.
Furthermore, due to the chance for overflow, we convert some of the
statistics class members into signed integers - overflow on unsigned
integers is undefined according to the C++ standard.